has Authenticated before allowing login
return 403 otherwise
