+++ clients/mockStoreClient.go
@@ -44,9 +44,11 @@ func (d MockStoreClient) FindUsers(user *models.User) (found []*models.User, err
 
 	if user.Name != "" {
 		found, _ := models.NewUser(&models.UserDetail{Name: user.Name, Pw: "123youknoWm3", Emails: []string{}}, d.salt)
+		found.Authenticated = true
 
 		return []*models.User{found}, nil
 	}
+	user.Authenticated = true
 
 	return []*models.User{user}, nil
 
@@ -61,13 +63,16 @@ func (d MockStoreClient) FindUser(user *models.User) (found *models.User, err er
 
 	if d.returnDifferent {
 		other, _ := models.NewUser(&models.UserDetail{Name: "Some One Else", Pw: "123youknoWm3", Emails: []string{}}, d.salt)
+		other.Authenticated = true
 		return other, nil
 	}
 
 	if user.Name != "" {
 		found, _ := models.NewUser(&models.UserDetail{Name: user.Name, Pw: "123youknoWm3", Emails: []string{}}, d.salt)
+		found.Authenticated = true
 		return found, nil
 	}
+	user.Authenticated = true
 	return user, nil
 }
